expire sessions if expire_after is given
We store the expiry date with the session. When an expired session is requested we destroy it and return a new session instead. In order to keep the database from growing with expired sessoins that have never been requested after they expired we'll also need a cleanup task to run in parallel.