Version 0.2.2 - catch connection failures
If we can't connect to the couch return false on attempts to store the session.
This way the application does not crash.
It also keeps the cookie middleware from storing useless cookies and the
session store will not try to retrieve the session from the couch on the next
request.