refactor: Session::Document < CouchRest::Document
It also stores the database information now.

Database info was not stored properly before leading to this issue
on travis:
NoMethodError: undefined method `get' for "sessions":String

The database methods in Store were conflicting with the ones
included from CouchRest::Model. Now we are mimicing the way
CouchRest::Model works.